--- libs/entitlement/qa/entitlement-allowed/src/main/java/module-info.java
@@ -1,15 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the "Elastic License
- * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
- * Public License v 1"; you may not use this file except in compliance with, at
- * your election, the "Elastic License 2.0", the "GNU Affero General Public
- * License v3.0 only", or the "Server Side Public License, v 1".
- */
-
-module org.elasticsearch.entitlement.qa.allowed {
-    requires org.elasticsearch.server;
-    requires org.elasticsearch.base;
-    requires org.elasticsearch.logging;
-    requires org.elasticsearch.entitlement.qa.common;
-}
