--- libs/entitlement/qa/entitlement-allowed/src/main/plugin-metadata/entitlement-policy.yaml
@@ -1,10 +1,0 @@
-org.elasticsearch.entitlement.qa.common:
-  - create_class_loader
-  - set_https_connection_properties
-  - inbound_network
-  - outbound_network
-  - write_system_properties:
-      properties:
-        - es.entitlements.checkSetSystemProperty
-        - es.entitlements.checkClearSystemProperty
-  - load_native_libraries
