--- libs/entitlement/qa/entitlement-denied-nonmodular/src/main/plugin-metadata/entitlement-policy.yaml
@@ -1,5 +1,0 @@
-ALL-UNNAMED:
-  - write_system_properties:
-      properties:
-        # entitlement itself not sufficient, also no wildcard support
-        - "*"
