--- libs/entitlement/qa/entitlement-denied/build.gradle
@@ -1,25 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the "Elastic License
- * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
- * Public License v 1"; you may not use this file except in compliance with, at
- * your election, the "Elastic License 2.0", the "GNU Affero General Public
- * License v3.0 only", or the "Server Side Public License, v 1".
- */
-
-apply plugin: 'elasticsearch.base-internal-es-plugin'
-
-esplugin {
-  name = 'entitlement-denied'
-  description = 'A test module that invokes non-granted entitlement and triggers exceptions'
-  classname = 'org.elasticsearch.entitlement.qa.EntitlementDeniedPlugin'
-}
-
-dependencies {
-  implementation project(':libs:entitlement:qa:common')
-}
-
-tasks.named("javadoc").configure {
-  enabled = false
-}
-
