--- libs/entitlement/qa/entitlement-denied/src/main/java/module-info.java
@@ -1,6 +1,0 @@
-module org.elasticsearch.entitlement.qa.denied {
-    requires org.elasticsearch.server;
-    requires org.elasticsearch.base;
-    requires org.apache.logging.log4j;
-    requires org.elasticsearch.entitlement.qa.common;
-}
