--- libs/entitlement/qa/entitlement-denied/src/main/plugin-metadata/entitlement-policy.yaml
@@ -1,5 +1,0 @@
-org.elasticsearch.entitlement.qa.common:
-  - write_system_properties:
-      properties:
-        # entitlement itself not sufficient, also no wildcard support
-        - "*"
